Move test to tests/python/relay/test_op_level2.py
